requirements.txt
pip~=23.1.2
attrs~=23.1.0
setuptools~=67.8.0
twitchio~=2.6.0
python-dotenv~=1.0.0
requests~=2.31.0